requirements.dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.dev.txt
#
--requirement requirements.txt

attrs==19.3.0             # via pytest
filelock==3.0.12          # via tox, virtualenv
importlib-metadata==1.7.0  # via pluggy, pytest, tox, virtualenv
iniconfig==1.0.1          # via pytest
more-itertools==8.4.0     # via pytest
packaging==20.4           # via pytest, tox
pathlib2==2.3.5           # via pytest
pluggy==0.13.1            # via pytest, tox
py==1.9.0                 # via pytest, tox
pyparsing==2.4.7          # via packaging
pytest==6.0.1             # via sof_wrapper (setup.py)
pytest-mock==3.3.0        # via pytest-mock (setup.py)
pytest-redis==2.0.0       # via pytest-redis (setup.py)
requests-mock==1.8.0      # via requests-mock
toml==0.10.1              # via pytest, tox
tox==3.19.0               # via sof_wrapper (setup.py)
virtualenv==20.0.30       # via tox
zipp==1.2.0               # via importlib-metadata
